requirements_frontend.txt
streamlit==1.44.1
requests==2.32.3
pillow==11.1.0
pandas==2.2.3
matplotlib==3.10.0
numpy==2.2.3
